refactor: split token_limiter.py into focused modules (#140)
Decompose 605-line token_limiter.py into four files respecting the
atomic design hierarchy and 150-line atom limit:

- lean/token_types.py: ContentType enum, TokenEstimate/TruncationConfig/
  TruncationResult dataclasses, _safe_json_serializer (~75 lines)
- lean/token_estimation.py: TokenEstimator class + ratio constants (~75 lines)
- lean/content_truncation.py: ContentTruncator class (~235 lines)
- lean/token_limiter.py: MCPTokenLimiter + apply_token_limits + full
  re-exports for backward compatibility (~175 lines)

All existing callers (interface.py, response_offloader.py, __init__.py,
tests) continue to import from lean.token_limiter without changes.
